+++ CHANGELOG.md
@@ -7,6 +7,7 @@ All notable changes to this project will be documented in this file.
 ### Added
 
 - Added a new script to automatically generate `changelog.html` from `CHANGELOG.md`, streamlining the build process.
+- Implemented automated compression of the `dist` folder, streamlining the release process.
 
 ## [1.4.2] - 2025-09-02